[Fitting] Move ComputeObject to dedicated header
